requirements.txt
gradio==6.1.0
gradio_client==2.0.1
h5py==3.15.1
ipykernel==6.17.0
ipython==8.6.0
jupyter_client==7.4.4
jupyter_core==4.11.2
keras==2.10.0
Keras-Preprocessing==1.1.2
Markdown==3.10
matplotlib==3.10.7
numpy==1.26.4
opencv-python==4.12.0.88
pandas==2.3.3
parso==0.8.3
pillow==12.0.0
PyYAML==6.0.3
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
tensorflow==2.10.1
